Implemented a Rauch-Tung-Striebel smoother. Experimented with smoothing tracks.
So far, in my limited playing around with one ensemble member for one
year (2005), it doesn't seem like the smoothing or filtering has much
beneficial effect. That is, the squared distance does not decrease
appreciable for many of the tracks over the course of the year. KF/RTS
smoother may be a dead end, but it was quite useful to implement both of
them.