Work on #6: Implement a Kalman Filter based tracker.
It stores each track that it alread knows about, then looks at each new
point and calculates the innovation (KalmanFilter.y). It then squares
and sums this to come up with a number that represents how likely it is
that the new point belongs to the track. The most likely point is chosen
for each track. It's not complete because each point can belong to more
than one track.