Close #10, close #14: ens. member to member track matching and composite track
I've used a combined approach for this: to save on computation a
continuously updated average track is calculated across all ensemble
members, that is then used to add tracks which match. This has the
wanted upshot of producing an averaged location at each timestep for all
the tracks across ensemble members, which I'm looking at now to see if
it will be useful. The idea will be to run an analysis using this new
average track and compare it with the other analyses I'm already
running. I need to get some results issues fixed first, and possibly
replace dist with a proper dist function.